+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+* Add support for weeks:
+
+  - Shorthands for `this_week`, `next_week` and `prev_week`
+  - `full_week?` and `one_week?`
+  - `next` and `previous` now handle weeks correctly
+  - Tests for biweekly calculations
+
+  *Edwin Vlieg*
+
 ## 0.1.0
 
 *   Initial import of the DateRange implementation from the internal implementation in the Moneybird code.